Added command parser
Introduced Picturebot class. It offers functionality to  send a picture as before and also to process commands sent in chat.
Commands can be specified inside the Picturebot class as a list of dictionaries.
Added command line argument to trigger command processing of picbot
Added command MakeMeHappy to send a picture on request